# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contoh API documentation",
        "description": "Contoh API doc menggunakan OpenAPI/Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "choirudin.emchagmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/api/apigallery": {
            "get": {
                "tags": [
                    "Get Gallery Portofolio"
                ],
                "summary": "Gallery Ananda Kusuma Putri",
                "description": "Data Gallery",
                "operationId": "GetGallery",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/storeGallery": {
            "post": {
                "tags": [
                    "upload gallery"
                ],
                "summary": "Create a new gallery",
                "description": "Create a new gallery with title, description, and optional image upload.",
                "operationId": "createGallery",
                "requestBody": {
                    "description": "Gallery details",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Gallery title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Gallery description",
                                        "type": "string"
                                    },
                                    "picture": {
                                        "description": "Image file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gallery created successfully"
                    }
                }
            }
        },
        "/api/greet": {
            "get": {
                "tags": [
                    "greeting"
                ],
                "summary": "Returns a Sample API response",
                "description": "A sample greeting to test out the API",
                "operationId": "greet",
                "parameters": [
                    {
                        "name": "firstname",
                        "in": "query",
                        "description": "nama depan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "description": "nama belakang",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        }
    }
}